Refactor main.js
