+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "keeptrack.space",
-  "version": "3.0.1",
+  "version": "3.0.2",
   "description": "Complex astrodynamics tools designed for non-engineers to make learning about orbital mechanics and satellite operations more accessible.",
   "author": "Theodore Kruczek",
   "contributors": [
