+++ src/js/main.js
@@ -32,12 +32,11 @@ import * as glm from '@app/js/lib/gl-matrix.js';
 import { atmosphere, earth, lineManager, moon, sun } from '@app/js/sceneManager/sceneManager.js';
 import { camPitch, camYaw, cameraManager, mouseX, mouseY, screenDragPoint } from '@app/js/cameraManager.js'
 import { db, settingsManager } from '@app/js/keeptrack-head.js';
+import { isselectedSatNegativeOne, selectSatManager } from '@app/js/selectSat.js'
+import { mathValue, watermarkedDataURL } from '@app/js/helpers.js';
 import { satCruncher, satScreenPositionArray, satSet } from '@app/js/satSet.js';
 import { ColorScheme } from '@app/js/color-scheme.js';
-import { SunCalc } from '@app/js/lib/suncalc.js';
-import { adviceList } from '@app/js/advice-module.js';
 import { groups } from '@app/js/groups.js';
-import { mathValue } from '@app/js/helpers.js';
 import { meshManager } from '@app/modules/meshManager.js';
 import { missileManager } from '@app/modules/missileManager.js';
 import { mobile } from '@app/js/mobile.js';
@@ -85,9 +84,7 @@ var clickedSat = 0;
 var maxPinchSize = Math.hypot(window.innerWidth, $(document).height());
 
 // Menu Variables
-var isEditSatMenuOpen = false;
 var isDOPMenuOpen = false;
-var isLookanglesMenuOpen = false;
 
 var pickFb, pickTex;
 var pMatrix = glm.mat4.create();
@@ -101,8 +98,6 @@ var mouseTimeout = null;
 var mouseSat = -1;
 
 var dragPoint = [0, 0, 0];
-var isCtrlPressed = false;
-var isShiftPressed = false;
 
 var isMouseMoving = false;
 var dragHasMoved = false;
@@ -111,6 +106,12 @@ var isPinching = false;
 var deltaPinchDistance = 0;
 var startPinchDistance = 0;
 var touchStartTime;
+
+var satLabelModeLastTime = 0;
+var isSatMiniBoxInUse = false;
+var labelCount;
+var hoverBoxOnSatMiniElements = [];
+var satHoverMiniDOM;
 
 var isShowNextPass = false;
 
@@ -126,17 +127,8 @@ let isShowDistance = true;
 // getEarthScreenPoint
 let rayOrigin, ptThru, rayDir, toCenterVec, dParallel, longDir, dPerp, dSubSurf, dSurf, ptSurf;
 
-// //////////////////////////////////////////////////////////////////////////
-// //////////////////////////////////////////////////////////////////////////
-//
-// Start Initialization
-//
-// //////////////////////////////////////////////////////////////////////////
-// //////////////////////////////////////////////////////////////////////////
-
 var initializeKeepTrack = () => {
-  mobile.checkMobileMode();
-  // initializeGpuManager();
+  mobile.checkMobileMode();  
   webGlInit();
   sun.init();
   earth.init();
@@ -154,11 +146,12 @@ var initializeKeepTrack = () => {
     setTimeout(function () {
       earth.loadHiRes();
       earth.loadHiResNight();
-      if (!settingsManager.offline && 'serviceWorker' in navigator) {
-        navigator.serviceWorker.register('./serviceWorker.js').then(function () {
-          console.debug(`[Service Worker] Installed!`);
-        });
-      }
+      // TOOD: Fix service worker with webpack
+      // if (!settingsManager.offline && 'serviceWorker' in navigator) {
+      //   navigator.serviceWorker.register('./serviceWorker.js').then(function () {
+      //     console.debug(`[Service Worker] Installed!`);
+      //   });
+      // }
     }, 0);
     if (!settingsManager.disableUI) {
       searchBox.init(satData);
@@ -212,6 +205,7 @@ var initializeKeepTrack = () => {
       satLinkManager.idToSatnum();
     })();
   });
+  selectSatManager.init(ColorScheme.group);
   if (settingsManager.isEnableRadarData) radarDataManager.init();
   dlManager.drawLoop(); // kick off the animationFrame()s
   if (!settingsManager.disableUI && !settingsManager.isDrawLess) {
@@ -225,55 +219,70 @@ var initializeKeepTrack = () => {
   }
 };
 
-// //////////////////////////////////////////////////////////////////////////
-// //////////////////////////////////////////////////////////////////////////
-//
-// End Initialization
-// Start Main Drawing loop
-//
-// //////////////////////////////////////////////////////////////////////////
-// //////////////////////////////////////////////////////////////////////////
-
 var dlManager = {};
-dlManager.drawDt = 0;
-dlManager.setDrawDt = (dt) => {
-  dlManager.drawDt = dt;
-};
-dlManager.drawLoopCallback = null;
-dlManager.setDrawLoopCallback = (cb) => {
-  dlManager.drawLoopCallback = cb;
-};
-dlManager.i = 0;
-dlManager.drawNow;
-dlManager.sat;
-dlManager.demoModeSatellite = 0;
-dlManager.demoModeLastTime = 0;
-dlManager.time = null;
-dlManager.dt = null;
-dlManager.t0 = 0;
-dlManager.isShowFPS = false;
-dlManager.fps = 0;
+// Setup dlManager
+{
+  dlManager.drawLoopCallback = null;
+  dlManager.setDrawLoopCallback = (cb) => {
+    dlManager.drawLoopCallback = cb;
+  };
+  dlManager.i = 0;  
+  dlManager.sat;
+  dlManager.demoModeSatellite = 0;
+  dlManager.demoModeLastTime = 0;
+  dlManager.dt = null;
+  dlManager.t0 = 0;
+  dlManager.isShowFPS = false;  
+}
 
 dlManager.drawLoop = (preciseDt) => {
-  // NOTE drawLoop has 7kb memory leak -- No Impact
+  // Restart the draw loop when ready to draw again
   requestAnimationFrame(dlManager.drawLoop);
-  dlManager.drawNow = Date.now();
-  dlManager.dt = preciseDt - dlManager.t0 || 0;
-  dlManager.fps = 1000 / dlManager.dt;
-  if (dlManager.isShowFPS) console.log(dlManager.fps);
-  dlManager.t0 = preciseDt;
-  if (dlManager.fps < 30) {
-    updateHoverDelayLimit = settingsManager.updateHoverDelayLimitBig;
-  } else if (dlManager.fps < 50) {
-    updateHoverDelayLimit = settingsManager.updateHoverDelayLimitSmall;
-  } else {
-    if (updateHoverDelayLimit > 1) --updateHoverDelayLimit;
-  }
 
-  dlManager.time = dlManager.drawNow;
-  timeManager.setNow(dlManager.drawNow);
+  // Record milliseconds since last drawLoop - default is 0
+  dlManager.dt = preciseDt - dlManager.t0 || 0;
+  // Record last Draw Time for Calculating Difference
+  dlManager.t0 = preciseDt;  
+
+  // Display it if that settings is enabled
+  if (dlManager.isShowFPS) console.log(1000 / timeManager.dt);
+
+  // Update official time for everyone else
+  timeManager.setNow(Date.now(), dlManager.dt, $('#datetime-input-tb'));
   
-  if (objectManager.selectedSat !== -1) {
+  // Calculate changes related to satellites objects
+  dlManager.satCalculate();  
+  
+  // Calculate camera changes needed since last draw
+  cameraManager.calculate(objectManager.selectedSat, dlManager.dt);      
+
+  // Missile oribts have to be updated every draw or they quickly become innacurate
+  dlManager.updateMissileOrbits();
+
+  // If in satellite view the orbit buffer needs to be updated every time
+  if (cameraManager.cameraType.current == cameraManager.cameraType.satellite) orbitManager.updateOrbitBuffer(objectManager.lastSelectedSat());
+
+  // Update Earth Direction
+  earth.update();
+
+  // Actually draw things now that math is done
+  drawScene();
+
+  // Update orbit currently being hovered over
+  dlManager.updateHover();
+
+  // callbacks at the end of the draw loop (this should be used more!)
+  dlManager.onDrawLoopComplete(dlManager.drawLoopCallback);
+  
+  // If Demo Mode do stuff
+  if (settingsManager.isDemoModeOn) _demoMode();  
+
+  // If in the process of taking a screenshot complete work for that
+  if (settingsManager.screenshotMode) dlManager.screenShot();
+};
+
+dlManager.satCalculate = () => {
+  if (objectManager.selectedSat !== -1) {  
     dlManager.sat = satSet.getSat(objectManager.selectedSat);
     objectManager.lastSelectedSat(objectManager.selectedSat);
     if (!dlManager.sat.static) {
@@ -298,7 +307,7 @@ dlManager.drawLoop = (preciseDt) => {
           meshManager.selectedSatPosition.z < dlManager.sat.position.z + 1.0
         ) {
           // Lerp to smooth difference between SGP4 and position+velocity
-          meshManager.lerpPosition(dlManager.sat.position, dlManager.drawDt);          
+          meshManager.lerpPosition(dlManager.sat.position, timeManager.drawDt);          
         } else {
           meshManager.updatePosition(dlManager.sat.position);
         }
@@ -306,30 +315,35 @@ dlManager.drawLoop = (preciseDt) => {
     }
     if (dlManager.sat.missile) orbitManager.setSelectOrbit(dlManager.sat.satId);
   }
-  
-  cameraManager.calculate(objectManager.selectedSat, dlManager.dt);      
+  if (objectManager.selectedSat !== dlManager.lastSelectedSat) {
+    selectSatManager.selectSat(objectManager.selectedSat);
+    if (objectManager.selectedSat !== -1) {
+      orbitManager.setSelectOrbit(objectManager.selectedSat);
+      if (objectManager.isSensorManagerLoaded && sensorManager.currentSensor.lat != null) {
+        lineManager.updateLineToSat(objectManager.selectedSat, satSet.getIdFromSensorName(sensorManager.currentSensor.name));
+      }
+      uiManager.updateMap();
+    }
+    if (objectManager.selectedSat === -1 && !isselectedSatNegativeOne) {
+      orbitManager.clearSelectOrbit();
+    }
+    if (objectManager.selectedSat !== -1 || (objectManager.selectedSat == -1 && !isselectedSatNegativeOne)) {
+      lineManager.drawWhenSelected();
+    }    
+    dlManager.lastSelectedSat = objectManager.selectedSat;
+  }
+};
 
+dlManager.updateMissileOrbits = () => {
   if (typeof missileManager != 'undefined' && missileManager.missileArray.length > 0) {
     for (dlManager.i = 0; dlManager.i < missileManager.missileArray.length; dlManager.i++) {
       orbitManager.updateOrbitBuffer(missileManager.missileArray[dlManager.i].id);
     }
-  }  
-
-  if (objectManager.selectedSat !== dlManager.lastSelectedSat) {
-    selectSat(objectManager.selectedSat);
-    if (objectManager.selectedSat !== -1 || (objectManager.selectedSat == -1 && !isselectedSatNegativeOne)) {
-      lineManager.drawWhenSelected();
-    }
-    dlManager.lastSelectedSat = objectManager.selectedSat;
   }
+};
 
-  dlManager.drawScene();
-  dlManager.updateHover();
-  dlManager.onDrawLoopComplete(dlManager.drawLoopCallback);
-  if (settingsManager.isDemoModeOn) _demoMode();  
-
-  if (settingsManager.screenshotMode) {
-    webGlInit();
+dlManager.screenShot = () => {
+  webGlInit();
     if (settingsManager.queuedScreenshot) return;
 
     setTimeout(function () {
@@ -345,7 +359,7 @@ dlManager.drawLoop = (preciseDt) => {
         copyrightStr = '';
       }
 
-      link.href = _watermarkedDataURL(canvasDOM[0], copyrightStr);
+      link.href = watermarkedDataURL(canvasDOM[0], copyrightStr);
       settingsManager.screenshotMode = false;
       settingsManager.queuedScreenshot = false;
       setTimeout(function () {
@@ -354,921 +368,32 @@ dlManager.drawLoop = (preciseDt) => {
       webGlInit();
     }, 200);
     settingsManager.queuedScreenshot = true;
-  }
-};
-
-// //////////////////////////////////////////////////////////////////////////
-// //////////////////////////////////////////////////////////////////////////
-//
-// End Main Drawing loop
-//
-// //////////////////////////////////////////////////////////////////////////
-// //////////////////////////////////////////////////////////////////////////
-
-/*
-function initializeGpuManager () {
-  const MINUTES_PER_DAY = 1440;
-  let gM = {};
-  gM.kern = {};
-  gM.transforms = {};
-  gM.settings = {};
-  gM.settings.steps = 16;
-  gM.settings.batchSize = 5;
-
-  gM.gpu = new GPU({
-    mode: settingsManager.gpujsMode
-  });
-
-  gM.transforms.satrec = (satrec) => {
-    let satrecArray = [];
-    let i = 0;
-    for (var key in satrec) {
-      if (satrec.hasOwnProperty(key)) {
-        if (key == 'method') {
-          let num = (satrec[key] == 'd') ? 1 : 0;
-          satrecArray.push(num);
-          i++;
-          continue;
-        }
-        if (key == 'init') {
-          let num = (satrec[key] == 'y') ? 1 : 0;
-          satrecArray.push(num);
-          i++;
-          continue;
-        }
-        if (key == 'operationmode') {
-          let num = (satrec[key] == 'i') ? 1 : 0;
-          satrecArray.push(num);
-          i++;
-          continue;
-        }
-        if (key == 'satnum') {
-          satrecArray.push(parseInt(satrec[key]));
-          i++;
-          continue;
-        }
-        // If None of the above
-        satrecArray.push(satrec[key]);
-        i++;
-        continue;
-      }
-    }
-    while (i < 100) {
-      satrecArray.push(0);
-      i++;
-    }
-    return satrecArray;
-  };
-
-  gM.getSat = (posArray, satnum) => {
-    let id = satSet.sccIndex[satnum];
-    let x = posArray[id*4+1][0][0];
-    let y = posArray[id*4+2][0][0];
-    let z = posArray[id*4+3][0][0];
-    console.log(`${x}, ${y}, ${z}`);
-  }
-
-  gM.satrecSetup = (numOfSats,offset, now) => {
-    now = (now) ? now : new Date();
-    offset = (offset) ? offset : 0;
-
-    var j = jday(
-        now.getUTCFullYear(),
-        now.getUTCMonth() + 1, // Note, this function requires months in range 1-12.
-        now.getUTCDate(),
-        now.getUTCHours(),
-        now.getUTCMinutes(),
-        now.getUTCSeconds()
-    );
-    j += now.getUTCMilliseconds() * 1.15741e-8; // days per millisecond
-
-    let satData = satSet.getSatData();
-    let gpuSatRecList = [];
-    let timeList = [];
-    let satrec;
-    for (var i = offset; i < offset + numOfSats; i++) {
-      satrec = satellite.twoline2satrec(satData[i].TLE1,satData[i].TLE2);
-      // if (satrec.satnum == "39208") debugger
-      gpuSatRecList.push(
-        gM.transforms.satrec(satrec)
-      );
-
-      timeList.push((j - satrec.jdsatepoch) * MINUTES_PER_DAY);
-    }
-    return [gpuSatRecList,timeList];
-  };
-
-  gM.funcs = {};
-  gM.funcs.sgp4 = (numOfSats,stepSize) => {
-    console.time("gM.funcs.sgp4");
-    let resArray = [];
-    let res;
-    let now = new Date();
-    for (var s = 0; s < numOfSats; s=s) {
-      let satStartIndex = s;
-      let satBatchSize = gM.settings.batchSize;
-      if (satStartIndex + gM.settings.batchSize > numOfSats) {
-        satBatchSize = numOfSats - satStartIndex;
-      }
-      s += satBatchSize;
-
-      [satrecList,tsinceList] = gM.satrecSetup(satBatchSize,satStartIndex, now);
-      gM.kern.sgp4.setOutput([4,gM.settings.steps,satrecList.length]);
-      // [j][t,x,y,z,xdot,ydot,zdot][id]
-      try {
-        res = gM.kern.sgp4(satrecList,tsinceList,stepSize);
-      } catch (e) {
-        console.log(s);
-        console.debug(e);
-      }
-      resArray = resArray.concat(res);
-    }
-
-    console.timeEnd("gM.funcs.sgp4");
-    return resArray;
-    // resArray[t,x,y,z][time][batch]
-  }
-  gM.funcs.sgp42 = (numOfSats,propLength) => {
-    // if (satrecList > 100) 'Max 100 satellites at a time!';
-    // if (propLength > 60*1440*2) 'Max Two Day Propagation';
-    // if (satrecList.length !== tsinceList.length) throw 'Parameters must be same length!';
-
-    let batches = Math.floor(numOfSats / 50);
-    let lastBatchSats = numOfSats % 50;
-    let tArray = [];
-    let xArray = [];
-    let yArray = [];
-    let zArray = [];
-    for (var b = 0; b <= batches; b++) {
-      let t = [];
-      let x = [];
-      let y = [];
-      let z = [];
-      if (b == batches) { // If Last Batch
-        if (lastBatchSats == 0) break;
-        [satrecList,tsinceList] = gM.satrecSetup(lastBatchSats,b*50);
-      } else {
-        [satrecList,tsinceList] = gM.satrecSetup(50,b*50);
-      }
-      gM.kern.sgp4.setOutput([satrecList.length,propLength,4]);
-      // [j][t,x,y,z,xdot,ydot,zdot][id]
-      // try {
-        [t, x, y, z] = gM.kern.sgp4(satrecList,tsinceList);
-      // } catch {
-        // debugger
-      // }
-      tArray = tArray.concat(t);
-      xArray = xArray.concat(x);
-      yArray = yArray.concat(y);
-      zArray = zArray.concat(z);
-    }
-
-    return xArray;
-  }
-
-  // SGP4 Of Satellites At Same Time Advanced by PropTime
-  gM.kern.sgp4 = gM.gpu.createKernel( function(satrecList,tsinceList,stepSize) {
-    let t = tsinceList[this.thread.z] + (stepSize * this.thread.y / 60); // per Second
-    if (this.thread.x == 0) return t;
-
-    let pi = 3.141592653589793;
-    let x2o3 = 2.0 / 3.0;
-    let j2 = 0.00108262998905;
-    let j3 = -0.00000253215306;
-    let j4 = -0.00000161098761;
-    let earthRadius = 6378.137; // in km
-    let twoPi = pi * 2;
-    let deg2rad = pi / 180.0;
-    let rad2deg = 180 / pi;
-    let minutesPerDay = 1440.0;
-    let mu = 398600.5; // in km3 / s2
-    let xke = 60.0 / Math.sqrt((earthRadius * earthRadius * earthRadius) / mu);
-    let tumin = 1.0 / xke;
-    let j3oj2 = j3 / j2;
-
-    var am = 0;
-    var axnl = 0;
-    var aynl = 0;
-    var betal = 0;
-    var cosim = 0;
-    var sinim = 0;
-    var cnod = 0;
-    var snod = 0;
-    var cos2u = 0;
-    var sin2u = 0;
-    var coseo1 = 0;
-    var sineo1 = 0;
-    var cosi = 0;
-    var sini = 0;
-    var cosip = 0;
-    var sinip = 0;
-    var cosisq = 0;
-    var cossu = 0;
-    var sinsu = 0;
-    var cosu = 0;
-    var sinu = 0;
-    var delm = 0;
-    var delomg = 0;
-    var dndt = 0;
-    var emsq = 0;
-    var ecose = 0;
-    var el2 = 0;
-    var eo1 = 0;
-    var esine = 0;
-    var argpm = 0;
-    var argpp = 0;
-    var pl = 0;
-    var rl = 0;
-    var rvdot = 0;
-    var rvdotl = 0;
-    var su = 0;
-    var t2 = 0;
-    var t3 = 0;
-    var t4 = 0;
-    var tc = 0;
-    var tem5 = 0;
-    var tempvar = 0;
-    var temp1 = 0;
-    var temp2 = 0;
-    var tempa = 0;
-    var tempe = 0;
-    var templ = 0;
-    var u = 0;
-    var ux = 0;
-    var uy = 0;
-    var uz = 0;
-    var vx = 0;
-    var vy = 0;
-    var vz = 0;
-    var inclm = 0;
-    var mm = 0;
-    var nm = 0;
-    var nodem = 0;
-    var xinc = 0;
-    var xincp = 0;
-    var xl = 0;
-    var xlm = 0;
-    var mp = 0;
-    var xmdf = 0;
-    var xmx = 0;
-    var xmy = 0;
-    var vnodedf = 0;
-    var xnode = 0;
-    var nodep = 0;
-    let mrt = 0.0;
-    var temp = 0;
-    var r0 = 0;
-    var r1 = 0;
-    var r2 = 0;
-    var v0 = 0;
-    var v1 = 0;
-    var v2 = 0;
-    var aycof = satrecList[this.thread.z][18];
-    var xlcof = satrecList[this.thread.z][40];
-    var con41 = satrecList[this.thread.z][19];
-    var x1mth2 = satrecList[this.thread.z][36];
-    var x7thm1 = satrecList[this.thread.z][37];
-
-    //  ------- update for secular gravity and atmospheric drag -----
-    xmdf = satrecList[this.thread.z][10] + satrecList[this.thread.z][38] * t;
-    let argpdf = satrecList[this.thread.z][9] + satrecList[this.thread.z][28] * t;
-    let nodedf = satrecList[this.thread.z][7] + satrecList[this.thread.z][39] * t;
-    argpm = argpdf;
-    mm = xmdf;
-    t2 = t * t;
-    nodem = nodedf + satrecList[this.thread.z][42] * t2;
-    tempa = 1.0 - satrecList[this.thread.z][20] * t;
-    tempe = satrecList[this.thread.z][5] * satrecList[this.thread.z][21] * t;
-    templ = satrecList[this.thread.z][32] * t2;
-
-    if (satrecList[this.thread.z][16] !== 1) {
-        delomg = satrecList[this.thread.z][29] * t;
-        //  sgp4fix use mutliply for speed instead of pow
-        var delmtemp = 1.0 + satrecList[this.thread.z][27] * Math.cos(xmdf);
-        delm =
-            satrecList[this.thread.z][41] *
-            (delmtemp * delmtemp * delmtemp - satrecList[this.thread.z][26]);
-        temp = delomg + delm;
-        mm = xmdf + temp;
-        argpm = argpdf - temp;
-        t3 = t2 * t;
-        t4 = t3 * t;
-        tempa =
-            tempa - satrecList[this.thread.z][23] * t2 - satrecList[this.thread.z][24] * t3 - satrecList[this.thread.z][25] * t4;
-        tempe =
-            tempe +
-            satrecList[this.thread.z][5] * satrecList[this.thread.z][22] * (Math.sin(mm) - satrecList[this.thread.z][30]);
-        templ =
-            templ +
-            satrecList[this.thread.z][33] * t3 +
-            t4 * (satrecList[this.thread.z][34] + t * satrecList[this.thread.z][35]);
-    }
-    nm = satrecList[this.thread.z][11];
-    var em = satrecList[this.thread.z][8];
-    inclm = satrecList[this.thread.z][6];
-    if (satrecList[this.thread.z][17] === 1) {
-        tc = t;
-
-        var irez = satrecList[this.thread.z][46];
-        var d2201 = satrecList[this.thread.z][47];
-        var d2211 = satrecList[this.thread.z][48];
-        var d3210 = satrecList[this.thread.z][49];
-        var d3222 = satrecList[this.thread.z][50];
-        var d4410 = satrecList[this.thread.z][51];
-        var d4422 = satrecList[this.thread.z][52];
-        var d5220 = satrecList[this.thread.z][53];
-        var d5232 = satrecList[this.thread.z][54];
-        var d5421 = satrecList[this.thread.z][55];
-        var d5433 = satrecList[this.thread.z][56];
-        var dedt = satrecList[this.thread.z][57];
-        var del1 = satrecList[this.thread.z][58];
-        var del2 = satrecList[this.thread.z][59];
-        var del3 = satrecList[this.thread.z][60];
-        var didt = satrecList[this.thread.z][61];
-        var dmdt = satrecList[this.thread.z][62];
-        var dnodt = satrecList[this.thread.z][63];
-        var domdt = satrecList[this.thread.z][64];
-        var argpo = satrecList[this.thread.z][9];
-        var argpdot = satrecList[this.thread.z][28];
-        var gsto = satrecList[this.thread.z][45];
-        var xfact = satrecList[this.thread.z][84];
-        var xlamo = satrecList[this.thread.z][95];
-        var no = satrecList[this.thread.z][11];
-        var atime = satrecList[this.thread.z][98];
-        var em = em;
-        var xli = satrecList[this.thread.z][99];
-        var xni = satrecList[this.thread.z][100];
-
-        var fasx2 = 0.13130908;
-        var fasx4 = 2.8843198;
-        var fasx6 = 0.37448087;
-        var g22 = 5.7686396;
-        var g32 = 0.95240898;
-        var g44 = 1.8014998;
-        var g52 = 1.050833;
-        var g54 = 4.4108898;
-        var rptim = 4.37526908801129966e-3; // equates to 7.29211514668855e-5 rad/sec
-        var stepp = 720.0;
-        var stepn = -720.0;
-        var step2 = 259200.0;
-
-        //  ----------- calculate deep space resonance effects -----------
-        dndt = 0.0;
-        var theta = (gsto + tc * rptim) % twoPi;
-        em = em + dedt * t;
-
-        inclm = inclm + didt * t;
-        argpm = argpm + domdt * t;
-        nodem = nodem + dnodt * t;
-        mm = mm + dmdt * t;
-
-        var ft = 0.0;
-        if (irez !== 0) {
-            //  sgp4fix streamline check
-            if (
-                atime === 0.0 ||
-                t * atime <= 0.0 ||
-                Math.abs(t) < Math.abs(atime)
-            ) {
-                atime = 0.0;
-                xni = no;
-                xli = xlamo;
-            }
-
-            // sgp4fix move check outside loop
-            var delt = 0;
-            if (t > 0.0) {
-                delt = stepp;
-            } else {
-                delt = stepn;
-            }
-            var iretn = 381; // added for do loop
-            var iret = 0; // added for loop
-            var xndt = 0;
-            var xldot = 0;
-            var xnddt = 0;
-            while (iretn === 381) {
-                //  ------------------- dot terms calculated -------------
-                //  ----------- near - synchronous resonance terms -------
-                if (irez !== 2) {
-                    xndt =
-                        del1 * Math.sin(xli - fasx2) +
-                        del2 * Math.sin(2.0 * (xli - fasx4)) +
-                        del3 * Math.sin(3.0 * (xli - fasx6));
-                    xldot = xni + xfact;
-                    xnddt =
-                        del1 * Math.cos(xli - fasx2) +
-                        2.0 * del2 * Math.cos(2.0 * (xli - fasx4)) +
-                        3.0 * del3 * Math.cos(3.0 * (xli - fasx6));
-                    xnddt = xnddt * xldot;
-                } else {
-                    // --------- near - half-day resonance terms --------
-                    var xomi = argpo + argpdot * atime;
-                    var x2omi = xomi + xomi;
-                    var x2li = xli + xli;
-                    xndt =
-                        d2201 * Math.sin(x2omi + xli - g22) +
-                        d2211 * Math.sin(xli - g22) +
-                        d3210 * Math.sin(xomi + xli - g32) +
-                        d3222 * Math.sin(-xomi + xli - g32) +
-                        d4410 * Math.sin(x2omi + x2li - g44) +
-                        d4422 * Math.sin(x2li - g44) +
-                        d5220 * Math.sin(xomi + xli - g52) +
-                        d5232 * Math.sin(-xomi + xli - g52) +
-                        d5421 * Math.sin(xomi + x2li - g54) +
-                        d5433 * Math.sin(-xomi + x2li - g54);
-                    xldot = xni + xfact;
-                    xnddt =
-                        d2201 * Math.cos(x2omi + xli - g22) +
-                        d2211 * Math.cos(xli - g22) +
-                        d3210 * Math.cos(xomi + xli - g32) +
-                        d3222 * Math.cos(-xomi + xli - g32) +
-                        d5220 * Math.cos(xomi + xli - g52) +
-                        d5232 * Math.cos(-xomi + xli - g52) +
-                        2.0 *
-                            (d4410 * Math.cos(x2omi + x2li - g44) +
-                                d4422 * Math.cos(x2li - g44) +
-                                d5421 * Math.cos(xomi + x2li - g54) +
-                                d5433 * Math.cos(-xomi + x2li - g54));
-                    xnddt = xnddt * xldot;
-                }
-                //  ----------------------- integrator -------------------
-                //  sgp4fix move end checks to end of routine
-                if (Math.abs(t - atime) >= stepp) {
-                    iret = 0;
-                    iretn = 381;
-                } else {
-                    ft = t - atime;
-                    iretn = 0;
-                }
-                if (iretn === 381) {
-                    xli = xli + xldot * delt + xndt * step2;
-                    xni = xni + xndt * delt + xnddt * step2;
-                    atime = atime + delt;
-                }
-            }
-            nm = xni + xndt * ft + xnddt * ft * ft * 0.5;
-            xl = xli + xldot * ft + xndt * ft * ft * 0.5;
-            if (irez !== 1) {
-                mm = xl - 2.0 * nodem + 2.0 * theta;
-                dndt = nm - no;
-            } else {
-                mm = xl - nodem - argpm + theta;
-                dndt = nm - no;
-            }
-            nm = no + dndt;
-        }
-    }
-
-    if (nm <= 0.0) {
-        //  printf("// error nm %f\n", nm);
-        // satrecList[this.thread.z].error = 2;
-        //  sgp4fix add return
-        return nm;
-    }
-    am = Math.pow(xke / nm, x2o3) * tempa * tempa;
-    nm = xke / Math.pow(am, 1.5);
-    em = em - tempe;
-
-    //  fix tolerance for error recognition
-    //  sgp4fix am is fixed from the previous nm check
-    if (em >= 1.0 || em < -0.001) {
-        // || (am < 0.95)
-        //  printf("// error em %f\n", em);
-        // satrecList[this.thread.z].error = 1;
-        //  sgp4fix to return if there is an error in eccentricity
-        return -401;
-    }
-    //  sgp4fix fix tolerance to avoid a divide by zero
-    if (em < 1.0e-6) {
-        em = 1.0e-6;
-    }
-    mm = mm + satrecList[this.thread.z][11] * templ;
-    xlm = mm + argpm + nodem;
-    emsq = em * em;
-    temp = 1.0 - emsq;
-
-    nodem = nodem % twoPi;
-    argpm = argpm % twoPi;
-    xlm = xlm % twoPi;
-    mm = (xlm - argpm - nodem) % twoPi;
-
-    //  ----------------- compute extra mean quantities -------------
-    sinim = Math.sin(inclm);
-    cosim = Math.cos(inclm);
-
-    //  -------------------- add lunar-solar periodics --------------
-    var ep = em;
-    xincp = inclm;
-    argpp = argpm;
-    nodep = nodem;
-    mp = mm;
-    sinip = sinim;
-    cosip = cosim;
-    if (satrecList[this.thread.z][17] === 1) {
-        var init = 0;
-        var opsmode = satrecList[this.thread.z][43];
-
-        // Copy satellite attributes into local variables for convenience
-        // and symmetry in writing formulae.
-        var alfdp = 0;
-        var betdp = 0;
-        var cosop = 0;
-        var sinop = 0;
-        var dalf = 0;
-        var dbet = 0;
-        var dls = 0;
-        var f2 = 0;
-        var f3 = 0;
-        var pe = 0;
-        var pgh = 0;
-        var ph = 0;
-        var pinc = 0;
-        var sel = 0;
-        var ses = 0;
-        var sghl = 0;
-        var vsghs = 0;
-        var vshs = 0;
-        var sil = 0;
-        var sinzf = 0;
-        var sis = 0;
-        var sll = 0;
-        var sls = 0;
-        var xls = 0;
-        var xnoh = 0;
-        var zf = 0;
-        var zm = 0;
-        var shll = 0;
-
-        var e3 = satrecList[this.thread.z][65];
-        var ee2 = satrecList[this.thread.z][66];
-        var peo = satrecList[this.thread.z][67];
-        var pgho = satrecList[this.thread.z][68];
-        var pho = satrecList[this.thread.z][69];
-        var pinco = satrecList[this.thread.z][70];
-        var plo = satrecList[this.thread.z][71];
-        var se2 = satrecList[this.thread.z][72];
-        var se3 = satrecList[this.thread.z][73];
-        var sgh2 = satrecList[this.thread.z][74];
-        var sgh3 = satrecList[this.thread.z][75];
-        var sgh4 = satrecList[this.thread.z][76];
-        var sh2 = satrecList[this.thread.z][77];
-        var sh3 = satrecList[this.thread.z][78];
-        var si2 = satrecList[this.thread.z][79];
-        var si3 = satrecList[this.thread.z][80];
-        var sl2 = satrecList[this.thread.z][81];
-        var sl3 = satrecList[this.thread.z][82];
-        var sl4 = satrecList[this.thread.z][83];
-        t = satrecList[this.thread.z][31];
-        var xgh2 = satrecList[this.thread.z][85];
-        var xgh3 = satrecList[this.thread.z][86];
-        var xgh4 = satrecList[this.thread.z][87];
-        var xh2 = satrecList[this.thread.z][88];
-        var xh3 = satrecList[this.thread.z][89];
-        var xi2 = satrecList[this.thread.z][90];
-        var xi3 = satrecList[this.thread.z][91];
-        var xl2 = satrecList[this.thread.z][92];
-        var xl3 = satrecList[this.thread.z][93];
-        var xl4 = satrecList[this.thread.z][94];
-        var zmol = satrecList[this.thread.z][96];
-        var zmos = satrecList[this.thread.z][97];
-
-        //  ---------------------- constants -----------------------------
-        var zns = 1.19459e-5;
-        var zes = 0.01675;
-        var znl = 1.5835218e-4;
-        var zel = 0.0549;
-
-        //  --------------- calculate time varying periodics -----------
-        zm = zmos + zns * t;
-        // be sure that the initial call has time set to zero
-        if (init === 1) {
-            zm = zmos;
-        }
-        zf = zm + 2.0 * zes * Math.sin(zm);
-        sinzf = Math.sin(zf);
-        f2 = 0.5 * sinzf * sinzf - 0.25;
-        f3 = -0.5 * sinzf * Math.cos(zf);
-        ses = se2 * f2 + se3 * f3;
-        sis = si2 * f2 + si3 * f3;
-        sls = sl2 * f2 + sl3 * f3 + sl4 * sinzf;
-        var sghs = sgh2 * f2 + sgh3 * f3 + sgh4 * sinzf;
-        var shs = sh2 * f2 + sh3 * f3;
-        zm = zmol + znl * t;
-        if (init === 1) {
-            zm = zmol;
-        }
-
-        zf = zm + 2.0 * zel * Math.sin(zm);
-        sinzf = Math.sin(zf);
-        f2 = 0.5 * sinzf * sinzf - 0.25;
-        f3 = -0.5 * sinzf * Math.cos(zf);
-        sel = ee2 * f2 + e3 * f3;
-        sil = xi2 * f2 + xi3 * f3;
-        sll = xl2 * f2 + xl3 * f3 + xl4 * sinzf;
-        sghl = xgh2 * f2 + xgh3 * f3 + xgh4 * sinzf;
-        shll = xh2 * f2 + xh3 * f3;
-        pe = ses + sel;
-        pinc = sis + sil;
-        pl = sls + sll;
-        pgh = sghs + sghl;
-        ph = shs + shll;
-
-        if (init === 0) {
-            pe = pe - peo;
-            pinc = pinc - pinco;
-            pl = pl - plo;
-            pgh = pgh - pgho;
-            ph = ph - pho;
-            xincp = xincp + pinc;
-            ep = ep + pe;
-            sinip = Math.sin(xincp);
-            cosip = Math.cos(xincp);
-
-            if (xincp >= 0.2) {
-                ph = ph / sinip;
-                pgh = pgh - cosip * ph;
-                argpp = argpp + pgh;
-                nodep = nodep + ph;
-                mp = mp + pl;
-            } else {
-                //  ---- apply periodics with lyddane modification ----
-                sinop = Math.sin(nodep);
-                cosop = Math.cos(nodep);
-                alfdp = sinip * sinop;
-                betdp = sinip * cosop;
-                dalf = ph * cosop + pinc * cosip * sinop;
-                dbet = -ph * sinop + pinc * cosip * cosop;
-                alfdp = alfdp + dalf;
-                betdp = betdp + dbet;
-                nodep = nodep % twoPi;
-                //  sgp4fix for afspc written intrinsic functions
-                //  nodep used without a trigonometric function ahead
-                if (nodep < 0.0 && opsmode === 0) {
-                    nodep = nodep + twoPi;
-                }
-                xls = mp + argpp + cosip * nodep;
-                dls = pl + pgh - pinc * nodep * sinip;
-                xls = xls + dls;
-                xnoh = nodep;
-                nodep = Math.atan2(alfdp, betdp);
-                //  sgp4fix for afspc written intrinsic functions
-                //  nodep used without a trigonometric function ahead
-                if (nodep < 0.0 && opsmode === 0) {
-                    nodep = nodep + twoPi;
-                }
-                if (Math.abs(xnoh - nodep) > pi) {
-                    if (nodep < xnoh) {
-                        nodep = nodep + twoPi;
-                    } else {
-                        nodep = nodep - twoPi;
-                    }
-                }
-                mp = mp + pl;
-                argpp = xls - mp - cosip * nodep;
-            }
-        }
-
-        if (xincp < 0.0) {
-            xincp = -xincp;
-            nodep = nodep + pi;
-            argpp = argpp - pi;
-        }
-        if (ep < 0.0 || ep > 1.0) {
-            // satrecList[this.thread.z].error = 3;
-            //  sgp4fix add return
-            return -402;
-        }
-    }
-    //  -------------------- long period periodics ------------------
-    if (satrecList[this.thread.z][17] === 1) {
-        sinip = Math.sin(xincp);
-        cosip = Math.cos(xincp);
-        aycof = -0.5 * j3oj2 * sinip;
-        //  sgp4fix for divide by zero for xincp = 180 deg
-        if (Math.abs(cosip + 1.0) > 1.5e-12) {
-            xlcof =
-                (-0.25 *
-                    j3oj2 *
-                    sinip *
-                    (3.0 + 5.0 * cosip)) /
-                (1.0 + cosip);
-        } else {
-            xlcof =
-                (-0.25 *
-                    j3oj2 *
-                    sinip *
-                    (3.0 + 5.0 * cosip)) /
-                1.5e-12;
-        }
-    }
-    axnl = ep * Math.cos(argpp);
-    temp = 1.0 / (am * (1.0 - ep * ep));
-    aynl = ep * Math.sin(argpp) + temp * aycof;
-    xl = mp + argpp + nodep + temp * xlcof * axnl;
-
-    //  --------------------- solve kepler's equation ---------------
-    u = (xl - nodep) % twoPi;
-    eo1 = u;
-    tem5 = 9999.9;
-    var ktr = 1;
-
-    //    sgp4fix for kepler iteration
-    //    the following iteration needs better limits on corrections
-    while (Math.abs(tem5) >= 1.0e-12 && ktr <= 10) {
-        sineo1 = Math.sin(eo1);
-        coseo1 = Math.cos(eo1);
-        tem5 = 1.0 - coseo1 * axnl - sineo1 * aynl;
-        tem5 = (u - aynl * coseo1 + axnl * sineo1 - eo1) / tem5;
-        if (tem5 > 0.95) {
-            tem5 = 0.95;
-        } else if (tem5 < -0.95) {
-            tem5 = -0.95;
-        }
-        eo1 = eo1 + tem5;
-        ktr = ktr + 1;
-    }
-    //  ------------- short period preliminary quantities -----------
-    ecose = axnl * coseo1 + aynl * sineo1;
-    esine = axnl * sineo1 - aynl * coseo1;
-    el2 = axnl * axnl + aynl * aynl;
-    pl = am * (1.0 - el2);
-    if (pl < 0.0) {
-        // satrecList[this.thread.z].error = 4;
-        //  sgp4fix add return
-        return -403;
-    } else {
-        rl = am * (1.0 - ecose);
-        var rdotl = (Math.sqrt(am) * esine) / rl;
-        rvdotl = Math.sqrt(pl) / rl;
-        betal = Math.sqrt(1.0 - el2);
-        temp = esine / (1.0 + betal);
-        sinu = (am / rl) * (sineo1 - aynl - axnl * temp);
-        cosu = (am / rl) * (coseo1 - axnl + aynl * temp);
-        su = Math.atan2(sinu, cosu);
-        sin2u = (cosu + cosu) * sinu;
-        cos2u = 1.0 - 2.0 * sinu * sinu;
-        temp = 1.0 / pl;
-        temp1 = 0.5 * j2 * temp;
-        temp2 = temp1 * temp;
-
-        //  -------------- update for short period periodics ------------
-        if (satrecList[this.thread.z][17] === 1) {
-            cosisq = cosip * cosip;
-            con41 = 3.0 * cosisq - 1.0;
-            x1mth2 = 1.0 - cosisq;
-            x7thm1 = 7.0 * cosisq - 1.0;
-        }
-        mrt =
-            rl * (1.0 - 1.5 * temp2 * betal * satrecList[this.thread.z][19]) +
-            0.5 * temp1 * satrecList[this.thread.z][36] * cos2u;
-        su = su - 0.25 * temp2 * satrecList[this.thread.z][37] * sin2u;
-        xnode = nodep + 1.5 * temp2 * cosip * sin2u;
-        xinc = xincp + 1.5 * temp2 * cosip * sinip * cos2u;
-        var mvt =
-            rdotl -
-            (nm * temp1 * satrecList[this.thread.z][36] * sin2u) / xke;
-        rvdot =
-            rvdotl +
-            (nm *
-                temp1 *
-                (satrecList[this.thread.z][36] * cos2u + 1.5 * satrecList[this.thread.z][19])) /
-                xke;
-
-        //  --------------------- orientation vectors -------------------
-        sinsu = Math.sin(su);
-        cossu = Math.cos(su);
-        snod = Math.sin(xnode);
-        cnod = Math.cos(xnode);
-        sini = Math.sin(xinc);
-        cosi = Math.cos(xinc);
-        xmx = -snod * cosi;
-        xmy = cnod * cosi;
-        ux = xmx * sinsu + cnod * cossu;
-        uy = xmy * sinsu + snod * cossu;
-        uz = sini * sinsu;
-        vx = xmx * cossu - cnod * sinsu;
-        vy = xmy * cossu - snod * sinsu;
-        vz = sini * cossu;
-
-        //  --------- position and velocity (in km and km/sec) ----------
-        r0 = mrt * ux * earthRadius;
-        r1 = mrt * uy * earthRadius;
-        r2 = mrt * uz * earthRadius;
-        v0 = (mvt * ux + rvdot * vx) * ((earthRadius * xke) / 60.0);
-        v1 = (mvt * uy + rvdot * vy) * ((earthRadius * xke) / 60.0);
-        v2 = (mvt * uz + rvdot * vz) * ((earthRadius * xke) / 60.0);
-    }
-    //  sgp4fix for decaying satellites
-    if (mrt < 1.0) {
-        // satrecList[this.thread.z].error = 6;
-        return -404;
-    }
-
-    // if (this.thread.y == 0) {
-      if (this.thread.x == 1) return r0;
-      if (this.thread.x == 2) return r1;
-      if (this.thread.x == 3) return r2;
-    // } else if (this.thread.y == 1) {
-      if (this.thread.x == 4) return v0;
-      if (this.thread.x == 5) return v1;
-      if (this.thread.x == 6) return v2;
-    // }
-  },{
-    dynamicOutput: true,
-    output: [10000,7,1]
-  })
-
-  function jday(year, mon, day, hr, minute, sec) {
-      'use strict';
-      return (
-          367.0 * year -
-          Math.floor(7 * (year + Math.floor((mon + 9) / 12.0)) * 0.25) +
-          Math.floor((275 * mon) / 9.0) +
-          day +
-          1721013.5 +
-          ((sec / 60.0 + minute) / 60.0 + hr) / 24.0 //  ut in days
-      );
-  };
-
-  // Export
-  window.gpuManager = gM;
 }
-*/
 
-var _watermarkedDataURL = (canvas, text) => {
-  var tempCanvas = document.createElement('canvas');
-  var tempCtx = tempCanvas.getContext('2d');
-  var cw, ch;
-  cw = tempCanvas.width = canvas.width;
-  ch = tempCanvas.height = canvas.height;
-  tempCtx.drawImage(canvas, 0, 0);
-  tempCtx.font = '24px nasalization';
-  var textWidth = tempCtx.measureText(text).width;
-  tempCtx.globalAlpha = 1.0;
-  tempCtx.fillStyle = 'white';
-  tempCtx.fillText(text, cw - textWidth - 30, ch - 30);
-  // tempCtx.fillStyle ='black'
-  // tempCtx.fillText(text,cw-textWidth-10+2,ch-20+2)
-  // just testing by adding tempCanvas to document
-  document.body.appendChild(tempCanvas);
-  let image = tempCanvas.toDataURL();
-  tempCanvas.parentNode.removeChild(tempCanvas);
-  return image;
-};
-
-dlManager.drawScene = () => {
+var drawScene = () => {
   // Drawing ColorIds for Picking Satellites
   gl.bindFramebuffer(gl.FRAMEBUFFER, gl.pickFb);
   gl.clearColor(0.0, 0.0, 0.0, 1.0);
   gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
-  gl.bindFramebuffer(gl.FRAMEBUFFER, null);
+  gl.bindFramebuffer(gl.FRAMEBUFFER, null);  
 
-  timeManager.setLastTime(timeManager.propTimeVar);
-  timeManager.updatePropTime();
-  timeManager.setSelectedDate(timeManager.propTimeVar);
+  dlManager.orbitsAbove();    
 
-  // Don't update the time input unless it is currently being viewed.
-  if (timeManager.lastTime - timeManager.propTimeVar < 300 && (settingsManager.isEditTime || !settingsManager.cruncherReady)) {
-    $('#datetime-input-tb').val(timeManager.selectedDate.toISOString().slice(0, 10) + ' ' + timeManager.selectedDate.toISOString().slice(11, 19));
-  }
-
-  if (cameraManager.cameraType.current == cameraManager.cameraType.astronomy || cameraManager.cameraType.current == cameraManager.cameraType.planetarium) {
-    dlManager.sensorPos = _calculateSensorPos(sensorManager.currentSensor);
-    _showOrbitsAbove();
-  } else {
-    dlManager.sensorPos = null;
-  }
-
-  // Hide satMiniBoxes When Not in Use
-  if (!settingsManager.isSatLabelModeOn || cameraManager.cameraType.current !== cameraManager.cameraType.planetarium) {
-    if (isSatMiniBoxInUse) {
-      satMiniBox.innerHTML = '';
-    }
-    isSatMiniBoxInUse = false;
-  }
-
-  cameraManager.update(dlManager.sat, dlManager.sensorPos);
-  if (cameraManager.cameraType.current == cameraManager.cameraType.satellite) orbitManager.updateOrbitBuffer(objectManager.lastSelectedSat());
+  cameraManager.update(dlManager.sat, dlManager.sensorPos);  
 
   gl.useProgram(gl.pickShaderProgram);
   gl.uniformMatrix4fv(gl.pickShaderProgram.uPMatrix, false, pMatrix);
   gl.uniformMatrix4fv(gl.pickShaderProgram.camMatrix, false, cameraManager.camMatrix);
 
-  // Draw Scene
   // gl.clearColor(0.0, 0.0, 0.0, 1.0);
   gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
-  
-  earth.update();
-  if (dlManager.fps > settingsManager.fpsThrottle1) {
+    
+  if (1000 / timeManager.dt > settingsManager.fpsThrottle1) {
     if (!settingsManager.enableLimitedUI && !settingsManager.isDrawLess) {
       sun.draw(pMatrix, cameraManager.camMatrix);
-      // Disabling Moon Until it is Fixed
       moon.draw(pMatrix, cameraManager.camMatrix);
     }
-  } else {
-    console.log(dlManager.fps);
-  }
+  } 
   if (!settingsManager.enableLimitedUI && !settingsManager.isDrawLess && cameraManager.cameraType.current !== cameraManager.cameraType.planetarium && cameraManager.cameraType.current !== cameraManager.cameraType.astronomy) {
     atmosphere.update();
     atmosphere.draw(pMatrix, cameraManager.camMatrix);
@@ -1284,7 +409,7 @@ dlManager.drawScene = () => {
     orbitManager.setSelectOrbit(objectManager.selectedSat);    
   }  
 
-  // Draw Satellite if Selected
+  // Draw Satellite Model if a satellite is selected and meshManager is loaded
   if (objectManager.selectedSat !== -1 && typeof meshManager != 'undefined' && meshManager.isReady) {
     let sat = dlManager.sat;
     // If 3D Models Available, then draw them on the screen
@@ -1448,99 +573,74 @@ dlManager.drawScene = () => {
       }
     }
   }
-
-  /* DEBUG - show the pickbuffer on a canvas */
-  // debugImageData.data = pickColorMap
-  /* debugImageData.data.set(pickColorMap)
-    debugContext.putImageData(debugImageData, 0, 0) */
 };
 
-var _calculateSensorPos = (sensor) => {
-  var now = timeManager.propTime();
-  var jday = (year, mon, day, hr, minute, sec) => 367.0 * year - Math.floor(7 * (year + Math.floor((mon + 9) / 12.0)) * 0.25) + Math.floor((275 * mon) / 9.0) + day + 1721013.5 + ((sec / 60.0 + minute) / 60.0 + hr) / 24.0; //  ut in days
-  var j = jday(
-    now.getUTCFullYear(),
-    now.getUTCMonth() + 1, // Note, this function requires months in range 1-12.
-    now.getUTCDate(),
-    now.getUTCHours(),
-    now.getUTCMinutes(),
-    now.getUTCSeconds()
-  );
-  j += now.getUTCMilliseconds() * 1.15741e-8; // days per millisecond
-  var gmst = satellite.gstime(j);
+dlManager.orbitsAbove = () => {
+  if (cameraManager.cameraType.current == cameraManager.cameraType.astronomy || cameraManager.cameraType.current == cameraManager.cameraType.planetarium) {
+    dlManager.sensorPos = satellite.calculateSensorPos(sensorManager.currentSensor);
+    // Previously called showOrbitsAbove();
+    if (!settingsManager.isSatLabelModeOn || cameraManager.cameraType.current !== cameraManager.cameraType.planetarium) {
+      if (isSatMiniBoxInUse) {
+        $('#sat-minibox').html('');
+      }
+      isSatMiniBoxInUse = false;
+      return;
+    }
+  
+    if (sensorManager.currentSensor.lat == null) return;
+    if (timeManager.now - satLabelModeLastTime < settingsManager.satLabelInterval) return;
+  
+    orbitManager.clearInViewOrbit();
+  
+    var sat;
+    labelCount = 0;
+    isHoverBoxVisible = true;
+  
+    hoverBoxOnSatMiniElements = document.getElementById('sat-minibox');
+    hoverBoxOnSatMiniElements.innerHTML = '';
+    for (var i = 0; i < satSet.orbitalSats && labelCount < settingsManager.maxLabels; i++) {
+      sat = satSet.getSatPosOnly(i);
+  
+      if (sat.static) continue;
+      if (sat.missile) continue;
+      if (sat.OT === 1 && ColorScheme.objectTypeFlags.payload === false) continue;
+      if (sat.OT === 2 && ColorScheme.objectTypeFlags.rocketBody === false) continue;
+      if (sat.OT === 3 && ColorScheme.objectTypeFlags.debris === false) continue;
+      if (sat.inview && ColorScheme.objectTypeFlags.inFOV === false) continue;
+  
+      satSet.getScreenCoords(i, pMatrix, cameraManager.camMatrix, sat.position);
+      if (satScreenPositionArray.error) continue;
+      if (typeof satScreenPositionArray.x == 'undefined' || typeof satScreenPositionArray.y == 'undefined') continue;
+      if (satScreenPositionArray.x > window.innerWidth || satScreenPositionArray.y > window.innerHeight) continue;
+  
+      // Draw Orbits
+      if (!settingsManager.isShowSatNameNotOrbit) {
+        orbitManager.addInViewOrbit(i);
+      }
+  
+      // Draw Sat Labels
+      // if (!settingsManager.enableHoverOverlay) continue
+      satHoverMiniDOM = document.createElement('div');
+      satHoverMiniDOM.id = 'sat-minibox-' + i;
+      satHoverMiniDOM.textContent = sat.SCC_NUM;
+      satHoverMiniDOM.setAttribute('style', 'display: block; position: absolute; left: ' + satScreenPositionArray.x + 10 + 'px; top: ' + satScreenPositionArray.y + 'px');
+      hoverBoxOnSatMiniElements.appendChild(satHoverMiniDOM);
+      labelCount++;
+    }
+    isSatMiniBoxInUse = true;
+    satLabelModeLastTime = timeManager.now;
+  } else {
+    dlManager.sensorPos = null;
+  }
 
-  var cosLat = Math.cos(sensor.lat * mathValue.DEG2RAD);
-  var sinLat = Math.sin(sensor.lat * mathValue.DEG2RAD);
-  var cosLon = Math.cos(sensor.long * mathValue.DEG2RAD + gmst);
-  var sinLon = Math.sin(sensor.long * mathValue.DEG2RAD + gmst);
-
-  let pos = {};
-  pos.x = (mathValue.RADIUS_OF_EARTH + mathValue.PLANETARIUM_DIST) * cosLat * cosLon;
-  pos.y = (mathValue.RADIUS_OF_EARTH + mathValue.PLANETARIUM_DIST) * cosLat * sinLon;
-  pos.z = (mathValue.RADIUS_OF_EARTH + mathValue.PLANETARIUM_DIST) * sinLat;
-  pos.gmst = gmst;
-  pos.lat = sensor.lat;
-  pos.long = sensor.long;
-  return pos;
-};
-
-var satLabelModeLastTime = 0;
-var isSatMiniBoxInUse = false;
-var labelCount;
-var hoverBoxOnSatMiniElements = [];
-var satHoverMiniDOM;
-var _showOrbitsAbove = () => {
+  // Hide satMiniBoxes When Not in Use
   if (!settingsManager.isSatLabelModeOn || cameraManager.cameraType.current !== cameraManager.cameraType.planetarium) {
     if (isSatMiniBoxInUse) {
-      $('#sat-minibox').html('');
+      satMiniBox.innerHTML = '';
     }
     isSatMiniBoxInUse = false;
-    return;
   }
-
-  if (sensorManager.currentSensor.lat == null) return;
-  if (dlManager.drawNow - satLabelModeLastTime < settingsManager.satLabelInterval) return;
-
-  orbitManager.clearInViewOrbit();
-
-  var sat;
-  labelCount = 0;
-  isHoverBoxVisible = true;
-
-  hoverBoxOnSatMiniElements = document.getElementById('sat-minibox');
-  hoverBoxOnSatMiniElements.innerHTML = '';
-  for (var i = 0; i < satSet.orbitalSats && labelCount < settingsManager.maxLabels; i++) {
-    sat = satSet.getSatPosOnly(i);
-
-    if (sat.static) continue;
-    if (sat.missile) continue;
-    if (sat.OT === 1 && ColorScheme.objectTypeFlags.payload === false) continue;
-    if (sat.OT === 2 && ColorScheme.objectTypeFlags.rocketBody === false) continue;
-    if (sat.OT === 3 && ColorScheme.objectTypeFlags.debris === false) continue;
-    if (sat.inview && ColorScheme.objectTypeFlags.inFOV === false) continue;
-
-    satSet.getScreenCoords(i, pMatrix, cameraManager.camMatrix, sat.position);
-    if (satScreenPositionArray.error) continue;
-    if (typeof satScreenPositionArray.x == 'undefined' || typeof satScreenPositionArray.y == 'undefined') continue;
-    if (satScreenPositionArray.x > window.innerWidth || satScreenPositionArray.y > window.innerHeight) continue;
-
-    // Draw Orbits
-    if (!settingsManager.isShowSatNameNotOrbit) {
-      orbitManager.addInViewOrbit(i);
-    }
-
-    // Draw Sat Labels
-    // if (!settingsManager.enableHoverOverlay) continue
-    satHoverMiniDOM = document.createElement('div');
-    satHoverMiniDOM.id = 'sat-minibox-' + i;
-    satHoverMiniDOM.textContent = sat.SCC_NUM;
-    satHoverMiniDOM.setAttribute('style', 'display: block; position: absolute; left: ' + satScreenPositionArray.x + 10 + 'px; top: ' + satScreenPositionArray.y + 'px');
-    hoverBoxOnSatMiniElements.appendChild(satHoverMiniDOM);
-    labelCount++;
-  }
-  isSatMiniBoxInUse = true;
-  satLabelModeLastTime = dlManager.drawNow;
-};
+}
 
 var currentSearchSats;
 dlManager.updateHover = () => {
@@ -1574,6 +674,14 @@ dlManager.updateHover = () => {
     // Earlier in the loop we decided how much to throttle updateHover
     // if we skip it this loop, we want to still draw the last thing
     // it was looking at
+
+    if (1000 / timeManager.dt < 30) {
+      updateHoverDelayLimit = settingsManager.updateHoverDelayLimitBig;
+    } else if (1000 / timeManager.dt < 50) {
+      updateHoverDelayLimit = settingsManager.updateHoverDelayLimitSmall;
+    } else {
+      if (updateHoverDelayLimit > 1) --updateHoverDelayLimit;
+    }
 
     if (++updateHoverDelay >= updateHoverDelayLimit) {
       updateHoverDelay = 0;
@@ -1768,9 +876,9 @@ dlManager.onDrawLoopComplete = (cb) => {
 var demoModeLastTime = 0;
 var _demoMode = () => {
   if (objectManager.isSensorManagerLoaded && sensorManager.currentSensor.lat == null) return;
-  if (dlManager.drawNow - demoModeLastTime < settingsManager.demoModeInterval) return;
+  if (timeManager.now - demoModeLastTime < settingsManager.demoModeInterval) return;
 
-  dlManager.demoModeLast = dlManager.drawNow;
+  dlManager.demoModeLast = timeManager.now;
 
   if (dlManager.demoModeSatellite === satSet.getSatData().length) dlManager.demoModeSatellite = 0;
   let satData = satSet.getSatData();
@@ -1794,29 +902,6 @@ var _demoMode = () => {
   }
 };
 
-// eslint-disable-next-line no-unused-vars
-var _fixDpi = (canvas, dpi) => {
-  //create a style object that returns width and height
-  let style = {
-    height() {
-      return +getComputedStyle(canvas).getPropertyValue('height').slice(0, -2);
-    },
-    width() {
-      return +getComputedStyle(canvas).getPropertyValue('width').slice(0, -2);
-    },
-  };
-  //set the correct attributes for a crystal clear image!
-  canvas.setAttribute('width', style.width() * dpi);
-  canvas.setAttribute('height', style.height() * dpi);
-};
-
-// Reinitialize the canvas on mobile rotation
-$(window).bind('orientationchange', function () {
-  console.log('rotate');
-  mobile.isRotationEvent = true;
-});
-
-cameraManager.webGlInit = () => webGlInit();
 var webGlInit = () => {
   db.log('webGlInit');
   let can = canvasDOM[0];
@@ -2014,11 +1099,13 @@ var _unProject = (mx, my) => {
   return [worldVec[0] / worldVec[3], worldVec[1] / worldVec[3], worldVec[2] / worldVec[3]];
 };
 var getSatIdFromCoord = (x, y) => {
-  // OPTIMIZE: Find a way to do this without using gl.readPixels!
+  // NOTE: gl.readPixels is a huge bottleneck
   gl.bindFramebuffer(gl.FRAMEBUFFER, gl.pickFb);
   gl.readPixels(x, gl.drawingBufferHeight - y, 1, 1, gl.RGBA, gl.UNSIGNED_BYTE, pickColorBuf);
   return ((pickColorBuf[2] << 16) | (pickColorBuf[1] << 8) | pickColorBuf[0]) - 1;
 };
+
+// Raycasting in getEarthScreenPoint would provide a lot of powerful (but slow) options later
 var getEarthScreenPoint = (x, y) => {
   rayOrigin = cameraManager.getCamPos();
   ptThru = _unProject(x, y);
@@ -2046,530 +1133,6 @@ var getEarthScreenPoint = (x, y) => {
   return ptSurf;
 };
 
-var isselectedSatNegativeOne = false;
-var selectSat = (satId) => {
-  db.log('selectSat');
-  db.log(`satId: ${satId}`, true);
-  var sat;
-  if (satId !== -1) {
-    cameraManager.rotateEarth(false);
-    sat = satSet.getSat(satId);
-    if (sat.type == 'Star') return;
-    if ((sat.active == false || typeof sat.active == 'undefined') && typeof sat.staticNum == 'undefined') return; // Non-Missile Non-Sensor Object
-  }
-  satSet.selectSat(satId);
-  cameraManager.setCamSnapMode(false);
-
-  if (satId === -1) {
-    if (settingsManager.currentColorScheme === ColorScheme.group || $('#search').val().length >= 3) {
-      // If group selected
-      $('#menu-sat-fov').removeClass('bmenu-item-disabled');
-    } else {
-      $('#menu-sat-fov').removeClass('bmenu-item-selected');
-      $('#menu-sat-fov').addClass('bmenu-item-disabled');
-      settingsManager.isSatOverflyModeOn = false;
-      satCruncher.postMessage({
-        isShowSatOverfly: 'reset',
-      });
-    }
-  }  
-
-  if (satId === -1 && !isselectedSatNegativeOne) {
-    cameraManager.fts2default();
-    isselectedSatNegativeOne = true;
-    $('#sat-infobox').fadeOut();
-    // $('#iss-stream').html('')
-    // $('#iss-stream-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000)
-    orbitManager.clearSelectOrbit();
-    // Remove Red Box
-    $('#menu-lookanglesmultisite').removeClass('bmenu-item-selected');
-    $('#menu-lookangles').removeClass('bmenu-item-selected');
-    $('#menu-editSat').removeClass('bmenu-item-selected');
-
-    $('#menu-map').removeClass('bmenu-item-selected');
-    $('#menu-newLaunch').removeClass('bmenu-item-selected');
-    $('#menu-breakup').removeClass('bmenu-item-selected');
-    $('#menu-customSensor').removeClass('bmenu-item-selected');
-    // Add Grey Out
-    $('#menu-lookanglesmultisite').addClass('bmenu-item-disabled');
-    $('#menu-lookangles').addClass('bmenu-item-disabled');
-    $('#menu-satview').addClass('bmenu-item-disabled');
-    $('#menu-editSat').addClass('bmenu-item-disabled');
-    $('#menu-map').addClass('bmenu-item-disabled');
-    $('#menu-newLaunch').addClass('bmenu-item-disabled');
-    $('#menu-breakup').addClass('bmenu-item-disabled');
-    // Remove Side Menus
-    // $('#lookanglesmultisite-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000)
-    // $('#lookangles-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000)
-    $('#editSat-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000);
-    $('#map-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000);
-    $('#newLaunch-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000);
-    $('#breakup-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000);
-    $('#customSensor-menu').effect('slide', { direction: 'left', mode: 'hide' }, 1000);
-
-    if ($('#search').val().length > 0) {
-      $('#search-results').attr('style', 'display: block; max-height:auto');
-    }
-
-    // Toggle the side menus as closed
-    isEditSatMenuOpen = false;
-    isLookanglesMenuOpen = false;
-    settingsManager.isMapMenuOpen = false;
-    sMM.isLookanglesMultiSiteMenuOpen(false);
-    sMM.isNewLaunchMenuOpen(false);
-    sMM.isBreakupMenuOpen(false);
-    sMM.isMissileMenuOpen(false);
-    uiManager.setCustomSensorMenuOpen(false);
-  } else if (satId !== -1) {
-    if (cameraManager.cameraType.current == cameraManager.cameraType.default) {
-      cameraManager.ecLastZoom = cameraManager.zoomLevel;
-      cameraManager.cameraType.set(cameraManager.cameraType.fixedToSat);
-    }
-    isselectedSatNegativeOne = false;
-    objectManager.setSelectedSat(satId);
-    sat = satSet.getSatExtraOnly(satId);
-    if (!sat) return;
-    if (sat.type == 'Star') {
-      return;
-    }
-    if (sat.static) {
-      if (typeof sat.staticNum == 'undefined') return;
-      adviceList.sensor();
-      sat = satSet.getSat(satId);
-      if (objectManager.isSensorManagerLoaded) sensorManager.setSensor(null, sat.staticNum); // Pass staticNum to identify which sensor the user clicked
-      if (objectManager.isSensorManagerLoaded) sensorManager.curSensorPositon = [sat.position.x, sat.position.y, sat.position.z];
-      objectManager.setSelectedSat(-1);
-      $('#menu-sensor-info').removeClass('bmenu-item-disabled');
-      $('#menu-fov-bubble').removeClass('bmenu-item-disabled');
-      $('#menu-surveillance').removeClass('bmenu-item-disabled');
-      $('#menu-planetarium').removeClass('bmenu-item-disabled');
-      $('#menu-astronomy').removeClass('bmenu-item-disabled');
-      if (objectManager.selectedSat !== -1) {
-        $('#menu-lookangles').removeClass('bmenu-item-disabled');
-      }
-      return;
-    }
-    cameraManager.camZoomSnappedOnSat(true);
-    cameraManager.camAngleSnappedOnSat(true);
-
-    orbitManager.setSelectOrbit(satId);
-
-    if (objectManager.isSensorManagerLoaded && sensorManager.currentSensor.lat != null) {
-      $('#menu-lookangles').removeClass('bmenu-item-disabled');
-    }
-
-    $('#menu-lookanglesmultisite').removeClass('bmenu-item-disabled');
-    $('#menu-satview').removeClass('bmenu-item-disabled');
-    $('#menu-editSat').removeClass('bmenu-item-disabled');
-    $('#menu-sat-fov').removeClass('bmenu-item-disabled');
-    $('#menu-map').removeClass('bmenu-item-disabled');
-    $('#menu-newLaunch').removeClass('bmenu-item-disabled');
-
-    if ($('#search-results').css('display') === 'block') {
-      if (window.innerWidth > 1000) {
-        if ($('#search').val().length > 0) {
-          $('#search-results').attr('style', 'display:block; max-height:27%');
-        }
-        if (cameraManager.cameraType.current !== cameraManager.cameraType.planetarium) {
-          // Unclear why this was needed...
-          // uiManager.legendMenuChange('default')
-        }
-      }
-    } else {
-      if (window.innerWidth > 1000) {
-        if ($('#search').val().length > 0) {
-          $('#search-results').attr('style', 'display:block; max-height:auto');
-        }
-        if (cameraManager.cameraType.current !== cameraManager.cameraType.planetarium) {
-          // Unclear why this was needed...
-          // uiManager.legendMenuChange('default')
-        }
-      }
-    }
-
-    if (!sat.missile) {
-      $('.sat-only-info').show();
-    } else {
-      $('.sat-only-info').hide();
-    }
-
-    $('#sat-infobox').fadeIn();
-    $('#sat-info-title').html(sat.ON);
-
-    if (sat.URL && sat.URL !== '') {
-      $('#sat-info-title').html("<a class='iframe' href='" + sat.URL + "'>" + sat.ON + '</a>');
-    }
-
-    $('#edit-satinfo-link').html("<a class='iframe' href='editor.htm?scc=" + sat.SCC_NUM + "&popup=true'>Edit Satellite Info</a>");
-
-    $('#sat-intl-des').html(sat.intlDes);
-    if (sat.OT === 'unknown') {
-      $('#sat-objnum').html(1 + sat.TLE2.substr(2, 7).toString());
-    } else {
-      //      $('#sat-objnum').html(sat.TLE2.substr(2,7))
-      $('#sat-objnum').html(sat.SCC_NUM);
-      if (settingsManager.isOfficialWebsite) {
-        // ga('send', 'event', 'Satellite', 'SCC: ' + sat.SCC_NUM, 'SCC Number');
-      }
-    }
-
-    var objtype;
-    if (sat.OT === 0) {
-      objtype = 'TBA';
-    }
-    if (sat.OT === 1) {
-      objtype = 'Payload';
-    }
-    if (sat.OT === 2) {
-      objtype = 'Rocket Body';
-    }
-    if (sat.OT === 3) {
-      objtype = 'Debris';
-    }
-    if (sat.OT === 4) {
-      if (settingsManager.offline) {
-        objtype = 'Special';
-      } else {
-        objtype = 'Amateur Sat';
-      }
-    }
-    if (sat.OT === 5) {
-      objtype = 'Measurement';
-    }
-    if (sat.OT === 6) {
-      objtype = 'Radar Track';
-    }
-    if (sat.OT === 7) {
-      objtype = 'Radar Object';
-    }
-    if (sat.missile) {
-      objtype = 'Ballistic Missile';
-    }
-    $('#sat-type').html(objtype);
-
-    // /////////////////////////////////////////////////////////////////////////
-    // Country Correlation Table
-    // /////////////////////////////////////////////////////////////////////////
-    var country;
-    country = objectManager.extractCountry(sat.C);
-    $('#sat-country').html(country);
-
-    // /////////////////////////////////////////////////////////////////////////
-    // Launch Site Correlation Table
-    // /////////////////////////////////////////////////////////////////////////
-    var site = [];
-    var missileLV;
-    var missileOrigin;
-    var satLvString;
-    if (sat.missile) {
-      site = sat.desc.split('(');
-      missileOrigin = site[0].substr(0, site[0].length - 1);
-      missileLV = sat.desc.split('(')[1].split(')')[0]; // Remove the () from the booster type
-
-      site.site = missileOrigin;
-      site.sitec = sat.C;
-    } else {
-      site = objectManager.extractLaunchSite(sat.LS);
-    }
-
-    $('#sat-site').html(site.site);
-    $('#sat-sitec').html(site.sitec);
-
-    if (settingsManager.isOfficialWebsite) {
-      // ga('send', 'event', 'Satellite', 'Country: ' + country, 'Country');
-    }
-    if (settingsManager.isOfficialWebsite) {
-      // ga('send', 'event', 'Satellite', 'Site: ' + site, 'Site');
-    }
-
-    // /////////////////////////////////////////////////////////////////////////
-    // Launch Vehicle Correlation Table
-    // /////////////////////////////////////////////////////////////////////////
-    if (sat.missile) {
-      sat.LV = missileLV;
-      $('#sat-vehicle').html(sat.LV);
-    } else {
-      $('#sat-vehicle').html(sat.LV); // Set to JSON record
-      if (sat.LV === 'U') {
-        $('#sat-vehicle').html('Unknown');
-      } // Replace with Unknown if necessary
-      satLvString = objectManager.extractLiftVehicle(sat.LV); // Replace with link if available
-      $('#sat-vehicle').html(satLvString);
-    }
-
-    // /////////////////////////////////////////////////////////////////////////
-    // RCS Correlation Table
-    // /////////////////////////////////////////////////////////////////////////
-    if (sat.R === null || typeof sat.R == 'undefined') {
-      $('#sat-rcs').html('Unknown');
-    } else {
-      var rcs;
-      if (sat.R < 0.1) {
-        rcs = 'Small';
-      }
-      if (sat.R >= 0.1) {
-        rcs = 'Medium';
-      }
-      if (sat.R > 1) {
-        rcs = 'Large';
-      }
-      $('#sat-rcs').html(rcs);
-      $('#sat-rcs').tooltip({ delay: 50, html: sat.R, position: 'left' });
-    }
-
-    if (!sat.missile) {
-      $('a.iframe').colorbox({
-        iframe: true,
-        width: '80%',
-        height: '80%',
-        fastIframe: false,
-        closeButton: false,
-      });
-      $('#sat-apogee').html(sat.apogee.toFixed(0) + ' km');
-      $('#sat-perigee').html(sat.perigee.toFixed(0) + ' km');
-      $('#sat-inclination').html((sat.inclination * mathValue.RAD2DEG).toFixed(2) + '°');
-      $('#sat-eccentricity').html(sat.eccentricity.toFixed(3));
-
-      $('#sat-period').html(sat.period.toFixed(2) + ' min');
-      $('#sat-period').tooltip({
-        delay: 50,
-        html: 'Mean Motion: ' + mathValue.MINUTES_PER_DAY / sat.period.toFixed(2),
-        position: 'left',
-      });
-
-      if (typeof sat.U != 'undefined' && sat.U != '') {
-        $('#sat-user').html(sat.U);
-      } else {
-        $('#sat-user').html('Unknown');
-      }
-      if (typeof sat.P != 'undefined' && sat.P != '') {
-        $('#sat-purpose').html(sat.P);
-      } else {
-        $('#sat-purpose').html('Unknown');
-      }
-      if (typeof sat.Con != 'undefined' && sat.Con != '') {
-        $('#sat-contractor').html(sat.Con);
-      } else {
-        $('#sat-contractor').html('Unknown');
-      }
-      if (typeof sat.LM != 'undefined' && sat.LM != '') {
-        $('#sat-lmass').html(sat.LM + ' kg');
-      } else {
-        $('#sat-lmass').html('Unknown');
-      }
-      if (typeof sat.DM != 'undefined' && sat.DM != '') {
-        $('#sat-dmass').html(sat.DM + ' kg');
-      } else {
-        $('#sat-dmass').html('Unknown');
-      }
-      if (typeof sat.Li != 'undefined' && sat.Li != '') {
-        $('#sat-life').html(sat.Li + ' yrs');
-      } else {
-        $('#sat-life').html('Unknown');
-      }
-      if (typeof sat.Pw != 'undefined' && sat.Pw != '') {
-        $('#sat-power').html(sat.Pw + ' w');
-      } else {
-        $('#sat-power').html('Unknown');
-      }
-      if (typeof sat.vmag != 'undefined' && sat.vmag != '') {
-        $('#sat-vmag').html(sat.vmag);
-      } else {
-        $('#sat-vmag').html('Unknown');
-      }
-      if (typeof sat.S1 != 'undefined' && sat.S1 != '') {
-        $('#sat-source1').html(`<a class="iframe" href="${sat.S1}">${sat.S1.split('//').splice(1)}</a>`);
-        $('#sat-source1w').show();
-      } else {
-        $('#sat-source1').html('Unknown');
-        $('#sat-source1w').hide();
-      }
-      if (typeof sat.S2 != 'undefined' && sat.S2 != '') {
-        $('#sat-source2').html(`<a class="iframe" href="${sat.S2}">${sat.S2.split('//').splice(1)}</a>`);
-        $('#sat-source2w').show();
-      } else {
-        $('#sat-source2').html('Unknown');
-        $('#sat-source2w').hide();
-      }
-      if (typeof sat.S3 != 'undefined' && sat.S3 != '') {
-        $('#sat-source3').html(`<a class="iframe" href="${sat.S3}">${sat.S3.split('//').splice(1)}</a>`);
-        $('#sat-source3w').show();
-      } else {
-        $('#sat-source3').html('Unknown');
-        $('#sat-source3w').hide();
-      }
-      if (typeof sat.S4 != 'undefined' && sat.S4 != '') {
-        $('#sat-source4').html(`<a class="iframe" href="${sat.S4}">${sat.S4.split('//').splice(1)}</a>`);
-        $('#sat-source4w').show();
-      } else {
-        $('#sat-source4').html('Unknown');
-        $('#sat-source4w').hide();
-      }
-      if (typeof sat.S5 != 'undefined' && sat.S5 != '') {
-        $('#sat-source5').html(`<a class="iframe" href="${sat.S5}">${sat.S5.split('//').splice(1)}</a>`);
-        $('#sat-source5w').show();
-      } else {
-        $('#sat-source5').html('Unknown');
-        $('#sat-source5w').hide();
-      }
-      if (typeof sat.S6 != 'undefined' && sat.S6 != '') {
-        $('#sat-source6').html(`<a class="iframe" href="${sat.S6}">${sat.S6.split('//').splice(1)}</a>`);
-        $('#sat-source6w').show();
-      } else {
-        $('#sat-source6').html('Unknown');
-        $('#sat-source6w').hide();
-      }
-      if (typeof sat.S7 != 'undefined' && sat.S7 != '') {
-        $('#sat-source7').html(`<a class="iframe" href="${sat.S7}">${sat.S7.split('//').splice(1)}</a>`);
-        $('#sat-source7w').show();
-      } else {
-        $('#sat-source7').html('Unknown');
-        $('#sat-source7w').hide();
-      }
-      if (typeof sat.URL != 'undefined' && sat.URL != '') {
-        $('#sat-source8').html(`<a class="iframe" href="${sat.URL}">${sat.URL.split('//').splice(1)}</a>`);
-        $('#sat-source8w').show();
-      } else {
-        $('#sat-source8').html('Unknown');
-        $('#sat-source8w').hide();
-      }
-      $('a.iframe').colorbox({
-        iframe: true,
-        width: '80%',
-        height: '80%',
-        fastIframe: false,
-        closeButton: false,
-      });
-
-      // TODO: Error checking on Iframe
-
-      var now = new Date();
-      var jday = timeManager.getDayOfYear(now);
-      now = now.getFullYear();
-      now = now.toString().substr(2, 2);
-      var daysold;
-      if (satSet.getSat(satId).TLE1.substr(18, 2) === now) {
-        daysold = jday - satSet.getSat(satId).TLE1.substr(20, 3);
-      } else {
-        daysold = jday - satSet.getSat(satId).TLE1.substr(20, 3) + satSet.getSat(satId).TLE1.substr(17, 2) * 365;
-      }
-      $('#sat-elset-age').html(daysold + ' Days');
-      $('#sat-elset-age').tooltip({
-        delay: 50,
-        html: 'Epoch Year: ' + sat.TLE1.substr(18, 2).toString() + ' Day: ' + sat.TLE1.substr(20, 8).toString(),
-        position: 'left',
-      });
-
-      if (!objectManager.isSensorManagerLoaded) {
-        $('#sat-sun').parent().hide();
-      } else {
-        now = new Date(timeManager.propRealTime + timeManager.propOffset);
-        var sunTime = SunCalc.getTimes(now, sensorManager.currentSensor.lat, sensorManager.currentSensor.long);
-        var satInSun = sat.isInSun;
-        // If No Sensor, then Ignore Sun Exclusion
-        if (sensorManager.currentSensor.lat == null) {
-          if (satInSun == 0) $('#sat-sun').html('No Sunlight');
-          if (satInSun == 1) $('#sat-sun').html('Limited Sunlight');
-          if (satInSun == 2) $('#sat-sun').html('Direct Sunlight');
-          // If Radar Selected, then Say the Sun Doesn't Matter
-        } else if (sensorManager.currentSensor.type !== 'Optical' && sensorManager.currentSensor.type !== 'Observer') {
-          $('#sat-sun').html('No Effect');
-          // If Dawn Dusk Can be Calculated then show if the satellite is in the sun
-        } else if (sunTime.dawn.getTime() - now > 0 || sunTime.dusk.getTime() - now < 0) {
-          if (satInSun == 0) $('#sat-sun').html('No Sunlight');
-          if (satInSun == 1) $('#sat-sun').html('Limited Sunlight');
-          if (satInSun == 2) $('#sat-sun').html('Direct Sunlight');
-          // If Optical Sesnor but Dawn Dusk Can't Be Calculated, then you are at a
-          // high latitude and we need to figure that out
-        } else if (sunTime.night != 'Invalid Date' && (sunTime.dawn == 'Invalid Date' || sunTime.dusk == 'Invalid Date')) {
-          if (satInSun == 0) $('#sat-sun').html('No Sunlight');
-          if (satInSun == 1) $('#sat-sun').html('Limited Sunlight');
-          if (satInSun == 2) $('#sat-sun').html('Direct Sunlight');
-        } else {
-          // Unless you are in sun exclusion
-          $('#sat-sun').html('Sun Exclusion');
-        }
-      }
-    }
-    if (objectManager.isSensorManagerLoaded && sensorManager.currentSensor.lat != null) {
-      if (isLookanglesMenuOpen) {
-        satellite.getlookangles(sat);
-      }
-      lineManager.updateLineToSat(satId, satSet.getIdFromSensorName(sensorManager.currentSensor.name));
-    }
-  }
-
-  objectManager.setSelectedSat(satId);
-
-  if (satId !== -1) {
-    if (typeof sat.TTP != 'undefined') {
-      $('#sat-ttp-wrapper').show();
-      $('#sat-ttp').html(sat.TTP);
-    } else {
-      $('#sat-ttp-wrapper').hide();
-    }
-    if (typeof sat.NOTES != 'undefined') {
-      $('#sat-notes-wrapper').show();
-      $('#sat-notes').html(sat.NOTES);
-    } else {
-      $('#sat-notes-wrapper').hide();
-    }
-    if (typeof sat.FMISSED != 'undefined') {
-      $('#sat-fmissed-wrapper').show();
-      $('#sat-fmissed').html(sat.FMISSED);
-    } else {
-      $('#sat-fmissed-wrapper').hide();
-    }
-    if (typeof sat.ORPO != 'undefined') {
-      $('#sat-oRPO-wrapper').show();
-      $('#sat-oRPO').html(sat.ORPO);
-    } else {
-      $('#sat-oRPO-wrapper').hide();
-    }
-    if (typeof sat.constellation != 'undefined') {
-      $('#sat-constellation-wrapper').show();
-      $('#sat-constellation').html(sat.constellation);
-    } else {
-      $('#sat-constellation-wrapper').hide();
-    }
-    if (typeof sat.maneuver != 'undefined') {
-      $('#sat-maneuver-wrapper').show();
-      $('#sat-maneuver').html(sat.maneuver);
-    } else {
-      $('#sat-maneuver-wrapper').hide();
-    }
-    if (typeof sat.associates != 'undefined') {
-      $('#sat-associates-wrapper').show();
-      $('#sat-associates').html(sat.associates);
-    } else {
-      $('#sat-associates-wrapper').hide();
-    }
-    uiManager.updateMap();
-
-    // ISS Stream Slows Down a Lot Of Computers
-    // if (sat.SCC_NUM === '25544') { // ISS is Selected
-    //   $('#iss-stream-menu').show()
-    //   $('#iss-stream').html('<iframe src="http://www.ustream.tv/embed/17074538?html5ui=1" allowfullscreen="true" webkitallowfullscreen="true" scrolling="no" frameborder="0" style="border: 0px none transparent"></iframe><iframe src="http://www.ustream.tv/embed/9408562?html5ui=1" allowfullscreen="true" webkitallowfullscreen="true" scrolling="no" frameborder="0" style="border: 0px none transparent"></iframe><br />' +
-    //                         '<iframe src="http://www.ustream.tv/embed/6540154?html5ui=1" allowfullscreen="true" webkitallowfullscreen="true" scrolling="no" frameborder="0" style="border: 0px none transparent"></iframe><iframe src="http://cdn.livestream.com/embed/spaceflightnow?layout=4&ampheight=340&ampwidth=560&ampautoplay=false" style="border:0outline:0" frameborder="0" scrolling="no"></iframe>')
-    // } else {
-    //   $('#iss-stream').html('')
-    //   $('#iss-stream-menu').hide()
-    // }
-  }
-};
-// eslint-disable-next-line no-unused-vars
-var enableSlowCPUMode = () => {
-  db.log('enableSlowCPUMode');
-  if (!settingsManager.cruncherReady) return;
-  settingsManager.isSlowCPUModeEnabled = true;
-  settingsManager.minimumSearchCharacters = 3;
-  settingsManager.satLabelInterval = 500;
-
-  satCruncher.postMessage({
-    isSlowCPUModeEnabled: true,
-  });
-};
-
 $(document).ready(function () {
   // Start the initialization before doing anything else. The webworkers and
   // textures needs to start loading as fast as possible.
@@ -2579,12 +1142,12 @@ $(document).ready(function () {
   // TODO: Migrate most things from UI to Here
   $(window).on({
     keydown: function (e) {
-      if (e.ctrlKey === true || e.metaKey === true) isCtrlPressed = true;
+      if (e.ctrlKey === true || e.metaKey === true) cameraManager.isCtrlPressed = true;
     },
   });
   $(window).on({
     keyup: function (e) {
-      if (e.ctrlKey === false && e.metaKey === false) isCtrlPressed = false;
+      if (e.ctrlKey === false && e.metaKey === false) cameraManager.isCtrlPressed = false;
     },
   });
 
@@ -2717,7 +1280,7 @@ $(document).ready(function () {
         if (evt.button === 1) {
           cameraManager.setLocalRotate(true);
           cameraManager.setLocalRotateStartPosition(cameraManager.localRotateCurrent);
-          if (isShiftPressed) {
+          if (cameraManager.isShiftPressed) {
             cameraManager.setLocalRotateRoll(true);
           cameraManager.setLocalRotateYaw(false);
           } else {
@@ -2728,10 +1291,10 @@ $(document).ready(function () {
         }
 
         // Right Mouse Button RMB
-        if (evt.button === 2 && (isShiftPressed || isCtrlPressed)) {
+        if (evt.button === 2 && (cameraManager.isShiftPressed || cameraManager.isCtrlPressed)) {
           cameraManager.isPanning = true;
           cameraManager.panStartPosition = cameraManager.panCurrent;
-          if (isShiftPressed) {
+          if (cameraManager.isShiftPressed) {
             cameraManager.isScreenPan = false;
             cameraManager.isWorldPan = true;
           } else {
@@ -2965,7 +1528,7 @@ $(document).ready(function () {
           }
           if (evt.button === 2) {
             // Right Mouse Button Clicked
-            if (!isCtrlPressed && !isShiftPressed) {
+            if (!cameraManager.isCtrlPressed && !cameraManager.isShiftPressed) {
               _openRmbMenu();
             }
           }
@@ -3487,7 +2050,7 @@ $(document).ready(function () {
           break;
         case 'edit-sat-rmb':
           objectManager.setSelectedSat(clickedSat);
-          if (!isEditSatMenuOpen) {
+          if (!sMM.isEditSatMenuOpen()) {
             uiManager.bottomIconPress({
               currentTarget: { id: 'menu-editSat' },
             });
@@ -3545,7 +2108,7 @@ $(document).ready(function () {
         case 'create-observer-rmb':
           $('#customSensor-menu').effect('slide', { direction: 'left', mode: 'show' }, 1000);
           $('#menu-customSensor').addClass('bmenu-item-selected');
-          uiManager.setCustomSensorMenuOpen(true);
+          sMM.setCustomSensorMenuOpen(true);
           $('#cs-lat').val(latLon.latitude);
           $('#cs-lon').val(latLon.longitude);
           $('#cs-hei').val(0);
@@ -3717,4 +2280,4 @@ $(document).ready(function () {
   })();
 });
 
-export { dlManager, gl };
+export { dlManager,  gl, webGlInit };
